Исправляем баг с работой IoTGpio

+++ include/classes/IoTGpio.h
@@ -6,12 +6,12 @@ class IoTGpio {
         IoTGpio(int pins);
         ~IoTGpio();
 
-        virtual void pinMode(uint8_t pin, uint8_t mode);
-        virtual void digitalWrite(uint8_t pin, uint8_t val);
-        virtual int digitalRead(uint8_t pin);
-        virtual int analogRead(uint8_t pin);
-        virtual void analogWrite(uint8_t pin, int val);
-        virtual void digitalInvert(uint8_t pin);
+        virtual void pinMode(int pin, uint8_t mode);
+        virtual void digitalWrite(int pin, uint8_t val);
+        virtual int digitalRead(int pin);
+        virtual int analogRead(int pin);
+        virtual void analogWrite(int pin, int val);
+        virtual void digitalInvert(int pin);
 
         int index;
         void regDriver(IoTGpio* newDriver);
